requirements.txt
flask==3.0.0
werkzeug==3.0.1
flask-limiter==3.5.0
flask-cors==4.0.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
limits==3.7.0